Add extra seeding check for Mt19937
For amusement, check the first variate of the newly-seeded sequence as
well as the 10_000th.  Both values in this unittest can be verified by
comparison to C++11 by compiling and running the following C++ program:

/****************************************************************/

int main ()
{
    static_assert(std::mt19937::default_seed == 5489,
                  "Default seed does not match Phobos!");
    std::mt19937 gen(std::mt19937::default_seed);
    std::cout << gen() << std::endl;
    for (int i = 0; i < 9998; ++i) {
        gen();
    }
    std::cout << gen() << std::endl;
}
/****************************************************************/

Note that the `for` loop in this example advances the generator 9998
times compared to the D unittest's `popFrontN(9999)` because the first
`gen()` call already advances the generator once.