Document undocumented overloads that take string
When looking at the signatures of std.file, it looks like
it only takes input ranges and specifically disallows strings.
The text of most functions say it can take strings, but the
proper overload isn't displayed at all.

This gave me a moment of confusion earlier today, so I decided to
just fix it with a bunch of little dittos. Now, the string overloads
should appear right with the non-string overloads.