Fix it so that isSomeString and isNarrowString are false for enums.
Having isSomeString and isNarrowString be true for enums whose base type
is a string type is a problem, because the enums will not work with the
same code that strings will, resulting either in compilation errors or
subtle bugs which result in variables of an enum type that contain
values that are not members of that enum type. As with other types that
implicitly convert to a string type, the enums need to be converted to
be used with the code that's designed for strings, which happens when
the function explicitly takes a string type or when the function takes
an array and is templated on the character/element type but does not
happen when the function is templated on the whole parameter type.

In addition to fixing isSomeString and isNarrowString for enums, these
changes catch and fix the fact that isNarrowString did not work
correctly with null - something that isSomeString did test for and
handle appropriately.