Debugger target for posix
It is desirable to be able to attach a debugger to the execution of
tests of phobos.
This patch add this feature.
By calling make -f posix.mak DEBUGGER=gdb std/XXXXX.debug
phobos and the module XXXXX will be compiled with debug symbols and
after its compilation the debugger gdb starts the compiled executable.
DEBUGGER=gdb can be omitted as it is the default debugger.