BitArray: Fix padding bits in length()
The function length() incorrectly changes the padding
bits of the BitArray when decreasing the length.

This patch makes the functionality of length() follow the
dynamic array semantic.

Also add a unit test to check the correctness of the state
of the bits when increasing or decreasing the length of the
BitArray.