CODEOWNERS: Add whitespace after # comments due to GitHub bug
According to the GitHub support:
We currently have a bug with CODEOWNERS files where commenting out a line
will stop the feature working entirely.
If you remove any lines with a # it should start to work correctly again.
Alternatively - since you have a lot of commented out lines in that file -
you can add a space after the # characters which should also fix the issue.