Fix issue 17801: DateTime.fromISOExtString no longer works with const strings.
The from*String functions in std.datetime did not have tests to make
sure that they worked with various character types or various levels of
constness, and some optimizations that were done to the from*String
functions on DateTime for 2.076 broke them for const strings. So, this
adds tests for all character types and levels of constness to each of
the from*String function in std.datetime, and fixes the functionality
for DateTime that then doesn't pass the tests.