Add std.concurrency.Generator overloads to accept Fiber's stack guard's size
As of 2.075 Fiber's stack protection pages are on by default, preventing
OS to merge mmaped ranges, so the total number of mmaped ranges is much
higher than before. In case there's no need for the stack overflow
protection and the high number of fibers, it may be wise to change this
setting, so the Generator's constructor overloads which support
this are added.

Fix issue 17914