[Trivial] improve documentation of popBackN
note that the popBackN doc is shared with popFrontN. This change moves the description of popBackN up, such that it is closer to where the reader will want to find it. Also adds more explicit mention of popFrontN, such that it is clear which one of the two is being described.