Deprecate public aliases for Regex, StaticRegex
As per the discussion at https://github.com/dlang/phobos/pull/5963, we
want the option to swap out the types for Regex, Capture, etc. This
is impossible if people are able to refer to Regex by name.
Furthermore, according to @wilzbach, aliases are not an option.
Therefore, this PR deprecates public uses of the Regex types to allow
us to switch out the concrete type later.