Fix Auto-Decoding Issue With DateTime From String Methods
DateTime's from string methods use countUntil in order to check for the seperator between
the date and the time strings. This is fine until the result is used to slice the original
string, which can cause incorrect results. This is due to the problem that countUntil gives
the number of code points until the Needle and not the number of code units, which is the
how it's sliced.

Normally this doesn't actually pose a problem because actual time strings only contain ASCII
characters, but this does fix two things

1. No more auto decoding in the function, so it's faster
2. Better error messages if there are non-ASCII characters in the string because the break-up is correct