Fix issue 15949: Make readText handle mismatched BOMs.
This adds a BOMException to std.encoding and makes readText check the
BOM. If there is one, and it doesn't match the requested type, then it
throws a BOMException. If the BOM matches and UTF validation fails or if
there is no BOM, and the size of the file mismatches with the requested
type (e.g. it's not divible by 2, so it can't be valid UTF-16), then a
UTFException is thrown.

readBOM was also added to read a BOM from a file. readText uses it, but
it's public so that anyone can use it.