logger update
this is a breaking change.
In order to make the logger work with allocator and the forward messages
without the need to gc allocation writeLogMsg had to go.
To test the design a new logger was added.
This logger is called AllocLogger and has an additional ctor argument,
an Allocator.

Additionally, the unit tests were simplified.