Add LibcurlPath version option
Adds possibility to specify absolute path to libcurl.
The TZDatabaseDir version option already does the same for the tzdata directory.
This commit enables the same fix for the path to libcurl. It's needed for Operating systems like NixOS without a search path for dynamic libs.
See https://issues.dlang.org/show_bug.cgi?id=15391 for more information.
Additionally I removed the Makefile part of the TZDatabaseDir solution
because the necessary compiler arguments can be easily given via DFLAGS, which is also compiler independent and doesn't use the /tmp dir which leads to problems on NixOS.