Mark etc.c.sqlite3 as deprecated
Bindings are for 3.10 which is outdated for some years now, the current
upstream version of sqlite3 is 3.28.

For up-to-date bindings, there are many external repositories in
code.dlang.org that provide them, such as d2sqlite or derelict.

If no one comes forward to maintain it, they will be removed.