logger: write thread id as part of log message
Previous implementation had `std.concurrency.Tid`, which is simply a
wrapper around `MessageBox` and is not related to Thread ID in anyway.
It cannot be used in `@safe` code.

Instead this patch uses `core.thread.ThreadID`. Using it requires `this`
and it also locks mutex for every call. This patch also caches the
result in a TLS to prevent unnecessary synchronization overhead.