[posix.mak]: Build @betterC extracted tests without -unittest
We don't want to compile -betterC test files with -unittest because this
will also compile the unit tests of imported files, and those unit tests
were not written in -betterC in mind.

PR https://github.com/dlang/tools/pull/386 allows us to not compile with
-unittest such files and this commit makes this change.

Also enable command output for the last 2 lines of the %.betterc rule is
it's quite helpful in order to see what's going on.