std.math: hide parameter names for common math functions
In anticipation of named parameters, this is an idea to establish a pattern to keep parameters out of the API when it makes sense.  Math functions that only take one parameter are a good example of functions where the name of the parameter is unhelpful to include in the function's "interface".