std.math: use 'radians' for trig functions instead of 'x'
Once named parameters are enabled, the parameter name 'x' doesn't convey any useful information to the caller, i.e.

sin(x: 90);

However, the name 'radians' would be useful for the client to include:

sin(radians: 90);

Now it's immediately apparent that the value is in 'radians', and we see that the caller likely meant PI/2 instead of 90.