Fix issue 20848: format should work with immutable(void)[]
The bug is that if `T == immutable(void)[]`, then `const(T)` does not
collapse to `const(void[])`, but is actually `const(immutable(void)[])`.
So a strict type equality check will fail.

However, `const(T)` *is* implicitly convertible to `const(void[])`,
which serves our purpose just as well. So just use that check instead.