UUID: Replace 'in ref' overloads with 'const scope ref'
'in' and 'in ref' overloads will conflict in the future if dlang/dmd#11000 is accepted.
Even if not, 'in' currently means something different based on a switch,
however the intention of this code is obviously the full 'const scope'.
In the some places, 'scope' was omitted as it is meaningless on value types.