std.mmfile: correct the documented type of thrown exceptions
MmFile.this claims that it throws std.file.FileException.
However, that never happens in practice:
  * On POSIX, ErrnoException is thrown (due to use of errnoEnforce).
  * On Windows, WindowsException is thrown (due to use of wenforce).

This change updates the documentation to reflect real types of
exceptions thrown by the constructors.

Another alternative is to change the MmFile itself to throw
FileException as it claims to do, but this will break backward
compatibility with existing programs.