std.range: fix function 'save' to only be implemented if the Range is has a valid 'save' method
The 'retro' Range only does one thing, iterates a Range back to front. This
follows the same semantics as a 'foreach_reverse'. However, with 'foreach_reverse'
it accepts Ranges that have valid 'popBack', 'back', and 'empty' functions. The
'retro' Range, on the other hand, forces the user to implement 'front',
'popFront', and 'save' functions as well because it follows the Range hierarchy.
For a Range that only iterates back to front, accomplishing the same as the
'foreach_reverse', it should behave the same way.

Fixes Issue #22272

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>