Fix a potential memory leak, typos
On line 2741, `if (nBytesNeeded < maxStack)` will force memory allocation in the case when `nBytesNeeded == maxStack` , even though it's not technically required, because the stack buffer can hold those bytes.
On line 2762 though, the check `if (nBytesNeeded > maxStack)` is slightly wrong then. If `nBytesNeeded == maxStack`, it would cause a memory leak.
The fix is to not allocate if they're equal.